% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run a Shiny application presenting the "markdownInput" package.}
\usage{
runExample()
}
\description{
Run a Shiny application presenting the "markdownInput" package.
}
